+++ CHANGELOG.md
@@ -24,7 +24,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Create empty project [#42](https://github.com/IN-CORE/incore-studio/issues/42)
 - Modal that adds resources from service to project [#40](https://github.com/IN-CORE/incore-studio/issues/40)
 - Create Execution component for displaying configurations and results [#5](https://github.com/IN-CORE/incore-studio/issues/5)
-
+- Display status of each execution steps [#39](https://github.com/IN-CORE/incore-studio/issues/39)
+-
 ### Changed
 
 - Navbar file structure. [#37](https://github.com/IN-CORE/incore-studio/issues/37)